extracted parts of pkg/gameservers that deal with SDK server into pkg/sdkserver
