+++ pkg/gameservers/controller_test.go
@@ -1106,3 +1106,14 @@ func newFakeController() (*Controller, agtesting.Mocks) {
 	c.recorder = m.FakeRecorder
 	return c, m
 }
+
+func newSingleContainerSpec() v1alpha1.GameServerSpec {
+	return v1alpha1.GameServerSpec{
+		Ports: []v1alpha1.GameServerPort{{ContainerPort: 7777, HostPort: 9999, PortPolicy: v1alpha1.Static}},
+		Template: corev1.PodTemplateSpec{
+			Spec: corev1.PodSpec{
+				Containers: []corev1.Container{{Name: "container", Image: "container/image"}},
+			},
+		},
+	}
+}
